Adds x64 macro and post-build deploy events.
The x64 macro will enable code not to be compiled when 32-bit binary is
built. This avoids creating broken executable with
architecture-dependent code.

The post-build deploy events will copy the built files (be it release or
debug, 32 or 64-bit) to hardcoded paths that should contain ATS and ETS2
plugins (hopefully correctly choosing the built architecture/bitness.