Implements code to fetch correct tank capacity from the game.
As the telemetry callback doesn't provide reliable fuel tank capacity
value, this code uses the memory address allocated to the configuration
data to search throughout game memory for a structure containing the
actual tank capacity. The code relies in telemetry data passing the
correct value for adblue tank's capacity.

The way it was implemented, it _should_ work in future game releases as
long as the fuel capacity structure is not changed. Currently the same
pattern matches both ATS 1.47.2.0 and ETS2 1.47.1.2, and even for trucks
not affected by the issue (trucks with customizable fuel tanks), the
correct value is brought in.

When the information is not available, the game may freeze for a short
time as the code will search all memory vicinity for pointers to the
structure.

We search pointers to the structure instead of the structure itself
because the pointers are more abundant (over 10 around the telemetry
memory address reference) against 3 to 4 copies of the structure across
game memory.

This (hopefully) fixes #1.